Starting Basics
Health implemented.  Working on implementing damage/healing.  Big problem at the moment is damage over time and healing over time effects.  Specifically: need to be careful as Coroutines are not applicable (due to the in-game clock running on its own timer).  Therefore the in-game clock is what everything is managed around!