More testing with turns
Major bug is slowing progress.  I think it's time to experiment with the turn system for now and figure out a better way.  The bug is this: when the player's turn is over, the game locks up and I have to force close the program.

Chances are this is directly related to how I'm calling functions and it's likely caused from some sort of infinite loop.