Turns Implemented, not tested, and big change
Implemented player turns in full and tested.  Implemented enemy turns but did not test.  As I was writing the enemy turn logic I realized a turn manager script would be more ideal to attach to any object that must require a "turn" to do anything.  I have not begun a full rework of this yet but it's the next big task.