Turn Sequence Complete
Turn sequences are complete.  Both the player and the enemy can act properly.  Developing a TurnManager script was the saving grace here.  I also edited the logic in the Timer script to better indicate when to start the timer again.

Something to note: this ONLY works with ONE enemy right now however can be fixed with a bool array!