Begun Actions Logic
Beginning to tinker with the action logic in my game.  For once I'm using, practically speaking, some inheritance and it's related techniques (very specifically: overriding methods).

Some needs before I can continue this path of the logic: I need a health script and to attach that to each entity (for now just the player and the single enemy).  No need to necessarily do death logic or destruction at this point but to test the actions I would like to use damage for now.

Comments in code are detailed enough to understand where my headspace is.  Hoping to continue this over the weekend!

Side note: I'm impressed with how far I've come in a year as a programmer.  Smile more, dude, you are getting better!