ActionQueue problems
Got the attacks to work (for their first attempt, any subsequent use of the same attack fails).  This is a problem and I think it's related to the below.  Taking a break to reset thoughts and let the frustration subside for now.

Also had an issue where they wouldn't trigger at all.  This was due to me destroying the ScriptableObject related to the action in the Player script rather than after the action triggered.

Right now I need to figure out where and when to destroy the action object (whether it's damage or Action).

If it isn't clear I've only tested this with the Player.  Once I get it in for the Player implementation with the enemy will be easier.

PS: lots of comments to help with additional information and helping understand what things do or are intended to do!