Full reworked Characters
All actors (referred to as characters) have been replaced as children of the Character.cs file.  So the player and enemies now have an abstract method they call from.  This contains all the required methods of an actor.

Secondly, I am now having an issue with the player stunning the enemy.  In debug mode, it steps through the code fine.  But when running normal: fails and leaves the game in a perpetual "player turn" state.

Need to review how I did this with the Enemy script to stun the player, find differences.