Fixed the stun
Fixed the stun completely.

The big problem was how I was using the passed stun time (I was basing it off the old logic from the attack and heal scripts).  Basically if the value of the stun was higher than 0, it delayed it for that time THEN stunned for the time.  When it should stun for that time, every time.

I also added in the scheduled boolean to help track turns a little better.  Allowing delay of turns to be more clear.  As well as when spellcasting is added.  Spells will not declare the user's next turn until the cast is finished.  Therefore it is vital that I have a way to handle when an entity doesn't have a turn scheduled.

For now I think handling them LOCALLY (ie: the entity itself) is best.  I've begun this, but barely, with the Player script.  The next step is to create a very basic spell cast that does damage but anytime the caster takes damage it should delay the cast slightly. (Also note: change health, if negative, is the ONLY one to delay, over time effects should NOT delay a cast).