Began implementing Slow, but new Stun interaction
Seems the stun is now going way overboard.  One of two things is happening: stun is calling a slow on top of it and causing an obnoxiously long turn timer.  OR Stun is calling more than once.  Not sure, need to fix that before continuing with slow.

NOTE: Also need to check for stun so we don't extend stuns with a slow!